refactor #20: 이미지 슬라이더 컴포넌트 분리, 애니메이션 에러 해결
